requirements.txt
appdirs==1.4.4
attrs==19.3.0
black==19.10b0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
google-api-core==1.22.0
google-api-python-client==1.10.0
google-auth==1.19.2
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
oauthlib==3.1.0
pathspec==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyQt5==5.15.0
PyQt5-sip==12.8.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
stringcase==1.2.0
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.10
wcwidth==0.2.5
